split routes to separated routers
