+++ api/routers/__init__.py
@@ -1,0 +1,7 @@
+from .auth import auth_router
+from .news import news_router
+
+__all__ = [
+    auth_router,
+    news_router,
+]
