+++ app.py
@@ -1,53 +1,19 @@
 import logging
-from typing import Annotated
 
-from fastapi import FastAPI, Header, HTTPException, Body, Depends
+from fastapi import FastAPI, Depends
 
-import conf
-from api.dependencies.auth import ApiKeyAuth, JWTTokenScopeAuth
-from repository.enums.scope import Scope
-from repository.news import NewsRepository
-from repository.user import UserRepository
-from schema.api.auth import UserTokenRequest, UserTokenResponse, BaseTokenPayload
-from schema.db.user import UserSchema
-from schema.query.user import UserEmailQuery
-from services.auth.token import TokenService
+from api.dependencies.auth import ApiKeyAuth
+from api.routers import auth_router, news_router
 
 logger = logging.getLogger(__name__)
 
 app = FastAPI()
 
 
-@app.get("/")
-async def root(api_key: str = Header()):
-    if api_key != conf.settings.API_KEY.get_secret_value():
-        raise HTTPException(status_code=403, detail="Invalid API key")
-
+@app.get("/", dependencies=[Depends(ApiKeyAuth)])
+async def root():
     return {"message": "Hello World"}
 
 
-@app.post("/auth", response_model=UserTokenResponse, dependencies=[Depends(ApiKeyAuth)])
-async def auth(data: UserTokenRequest = Body(),
-               user_repository: UserRepository = Depends(UserRepository)):
-    db_query = UserEmailQuery(email=data.email)
-    user = user_repository.read(db_query)
-
-    if not user:
-        user = user_repository.create(query=UserSchema(
-            email=data.email,
-            full_name=data.full_name,
-        ))
-
-    token = TokenService().generate_token(
-        payload=BaseTokenPayload(id=user.id, **data.model_dump(exclude={"minutes"})),
-        ttl=data.minutes)
-
-    logger.info(f"Token generated for \"{data.model_dump_json()}\" request body")
-
-    return UserTokenResponse(token=token)
-
-
-@app.get("/news", )
-async def get_news(user: Annotated[UserSchema, Depends(JWTTokenScopeAuth(required_scope=[Scope.NEWS]))]):
-    logger.info(f"User {user.email} got news response")
-    return NewsRepository().read_all()
+app.include_router(auth_router)
+app.include_router(news_router)
